% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-methods.R
\name{autoplot-halfmoon}
\alias{autoplot-halfmoon}
\alias{autoplot.halfmoon_balance}
\alias{autoplot.halfmoon_ess}
\alias{autoplot.halfmoon_calibration}
\alias{autoplot.halfmoon_roc}
\alias{autoplot.halfmoon_auc}
\alias{autoplot.halfmoon_qq}
\title{Autoplot Methods for halfmoon Objects}
\usage{
\method{autoplot}{halfmoon_balance}(object, ...)

\method{autoplot}{halfmoon_ess}(object, ...)

\method{autoplot}{halfmoon_calibration}(object, ...)

\method{autoplot}{halfmoon_roc}(object, ...)

\method{autoplot}{halfmoon_auc}(object, ...)

\method{autoplot}{halfmoon_qq}(object, ...)
}
\arguments{
\item{object}{A halfmoon data object with appropriate class}

\item{...}{Additional arguments passed to the underlying plot_*() function}
}
\value{
A ggplot2 object
}
\description{
These methods provide automatic plot generation for halfmoon data objects
using ggplot2's autoplot interface. Each method dispatches to the appropriate
plot_*() function as follows:
}
\details{
\itemize{
\item \code{autoplot.halfmoon_balance} calls \code{\link[=plot_balance]{plot_balance()}}
\item \code{autoplot.halfmoon_ess} calls \code{\link[=plot_ess]{plot_ess()}}
\item \code{autoplot.halfmoon_calibration} calls \code{\link[=plot_model_calibration]{plot_model_calibration()}}
\item \code{autoplot.halfmoon_roc} calls \code{\link[=plot_model_roc_curve]{plot_model_roc_curve()}}
\item \code{autoplot.halfmoon_auc} calls \code{\link[=plot_model_auc]{plot_model_auc()}}
\item \code{autoplot.halfmoon_qq} calls \code{\link[=plot_qq]{plot_qq()}}
}
}
