% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ecdf.R
\name{geom_ecdf}
\alias{geom_ecdf}
\title{Calculate weighted and unweighted empirical cumulative distributions}
\usage{
geom_ecdf(
  mapping = NULL,
  data = NULL,
  geom = "step",
  position = "identity",
  ...,
  n = NULL,
  pad = TRUE,
  na.rm = FALSE,
  orientation = NA,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{geom}{The geometric object to use to display the data for this layer.
When using a \verb{stat_*()} function to construct a layer, the \code{geom} argument
can be used to override the default coupling between stats and geoms. The
\code{geom} argument accepts the following:
\itemize{
\item A \code{Geom} ggproto subclass, for example \code{GeomPoint}.
\item A string naming the geom. To give the geom as a string, strip the
function name of the \code{geom_} prefix. For example, to use \code{geom_point()},
give the geom as \code{"point"}.
\item For more information and other ways to specify the geom, see the
\link[ggplot2:layer_geoms]{layer geom} documentation.
}}

\item{position}{A position adjustment to use on the data for this layer. This
can be used in various ways, including to prevent overplotting and
improving the display. The \code{position} argument accepts the following:
\itemize{
\item The result of calling a position function, such as \code{position_jitter()}.
This method allows for passing extra arguments to the position.
\item A string naming the position adjustment. To give the position as a
string, strip the function name of the \code{position_} prefix. For example,
to use \code{position_jitter()}, give the position as \code{"jitter"}.
\item For more information and other ways to specify the position, see the
\link[ggplot2:layer_positions]{layer position} documentation.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}'s \code{params} argument. These
arguments broadly fall into one of 4 categories below. Notably, further
arguments to the \code{position} argument, or aesthetics that are required
can \emph{not} be passed through \code{...}. Unknown arguments that are not part
of the 4 categories below are ignored.
\itemize{
\item Static aesthetics that are not mapped to a scale, but are at a fixed
value and apply to the layer as a whole. For example, \code{colour = "red"}
or \code{linewidth = 3}. The geom's documentation has an \strong{Aesthetics}
section that lists the available options. The 'required' aesthetics
cannot be passed on to the \code{params}. Please note that while passing
unmapped aesthetics as vectors is technically possible, the order and
required length is not guaranteed to be parallel to the input data.
\item When constructing a layer using
a \verb{stat_*()} function, the \code{...} argument can be used to pass on
parameters to the \code{geom} part of the layer. An example of this is
\code{stat_density(geom = "area", outline.type = "both")}. The geom's
documentation lists which parameters it can accept.
\item Inversely, when constructing a layer using a
\verb{geom_*()} function, the \code{...} argument can be used to pass on parameters
to the \code{stat} part of the layer. An example of this is
\code{geom_area(stat = "density", adjust = 0.5)}. The stat's documentation
lists which parameters it can accept.
\item The \code{key_glyph} argument of \code{\link[ggplot2:layer]{layer()}} may also be passed on through
\code{...}. This can be one of the functions described as
\link[ggplot2:draw_key]{key glyphs}, to change the display of the layer in the legend.
}}

\item{n}{if NULL, do not interpolate. If not NULL, this is the number
of points to interpolate with.}

\item{pad}{If \code{TRUE}, pad the ecdf with additional points (-Inf, 0)
and (Inf, 1)}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{orientation}{The axis the curve runs along, \code{"x"} or \code{"y"}. Defaults to
\code{NA}, which reads the orientation from the aesthetics the layer is given.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display. To include legend keys for all levels, even
when no data exists, use \code{TRUE}.  If \code{NA}, all levels are shown in legend,
but unobserved levels are omitted.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:annotation_borders]{annotation_borders()}}.}
}
\value{
a geom
}
\description{
The empirical cumulative distribution function (ECDF) provides an alternative
visualization of distribution. \code{geom_ecdf()} is similar to
\code{\link[ggplot2:stat_ecdf]{ggplot2::stat_ecdf()}} but it can also calculate weighted ECDFs.
}
\details{
ECDF plots show the cumulative distribution function \eqn{F(x) = P(X \leq x)},
displaying what proportion of observations fall below each value. When comparing
treatment groups, overlapping ECDF curves indicate similar distributions and
thus good balance.

ECDF plots are closely related to quantile-quantile (QQ) plots (see \code{\link[=geom_qq2]{geom_qq2()}}).
While ECDF plots show \eqn{F(x)} for each group, QQ plots show the inverse relationship
by plotting \eqn{F_1^{-1}(p)} vs \eqn{F_2^{-1}(p)}. Both visualize the same distributional
information:
\itemize{
\item ECDF plots: Compare cumulative probabilities at each value
\item QQ plots: Compare values at each quantile
}

Choose ECDF plots when you want to see the full cumulative distribution or when
comparing multiple groups simultaneously. Choose QQ plots when you want to directly
compare two groups with an easy-to-interpret 45-degree reference line.

\code{geom_ecdf()} supports both orientations. Mapping the variable to \code{y}, or
passing \code{orientation = "y"}, computes the same weighted curve and draws it
across the panel instead of up it.
}
\section{Aesthetics}{
 In addition to the aesthetics for
\code{\link[ggplot2:stat_ecdf]{ggplot2::stat_ecdf()}}, \code{geom_ecdf()} also accepts: \itemize{ \item
weights }
}

\examples{
library(ggplot2)

ggplot(
  nhefs_weights,
  aes(x = smokeyrs, color = qsmk)
) +
  geom_ecdf(aes(weights = w_ato)) +
  xlab("Smoking Years") +
  ylab("Proportion <= x")

}
\seealso{
\itemize{
\item \code{\link[=geom_qq2]{geom_qq2()}} for an alternative visualization using quantile-quantile plots
\item \code{\link[ggplot2:stat_ecdf]{ggplot2::stat_ecdf()}} for the unweighted version
}

Other ggplot2 functions:
\code{\link[=geom_calibration]{geom_calibration()}},
\code{\link[=geom_mirror_density]{geom_mirror_density()}},
\code{\link[=geom_mirror_histogram]{geom_mirror_histogram()}},
\code{\link[=geom_qq2]{geom_qq2()}},
\code{\link[=geom_roc]{geom_roc()}}
}
\concept{ggplot2 functions}
