% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ci.R
\name{get_z_score}
\alias{get_z_score}
\title{Get z-score for confidence level}
\usage{
get_z_score(conf_level = 0.95)
}
\arguments{
\item{conf_level}{Confidence level (default 0.95)}
}
\value{
z-score for the given confidence level
}
\description{
Get z-score for confidence level
}
\keyword{internal}
