% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-documentation.R
\name{ggplot2_params}
\alias{ggplot2_params}
\title{Parameter Documentation for ggplot2 Functions}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot.}

\item{data}{The data to be displayed in this layer. If \code{NULL}, the default,
the data is inherited from the plot data as specified in the call to \code{ggplot()}.}

\item{stat}{The statistical transformation to use on the data for this layer.}

\item{position}{Position adjustment.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning.
If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{Logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than
combining with them.}

\item{...}{Other arguments passed on to layer().}
}
\description{
Parameter Documentation for ggplot2 Functions
}
\keyword{internal}
