% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-documentation.R
\name{treatment_param}
\alias{treatment_param}
\title{Parameter Documentation for Treatment Level}
\arguments{
\item{treatment_level}{The level of the outcome variable to consider as the
treatment/event. If \code{NULL} (default), uses the last level for factors or
the maximum value for numeric variables.}
}
\description{
Parameter Documentation for Treatment Level
}
\keyword{internal}
