% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ess_header.R
\name{add_ess_header}
\alias{add_ess_header}
\title{Add ESS Header to gtsummary tbl}
\usage{
add_ess_header(x, header = "**{level}**  \\nN = {round(n)}")
}
\arguments{
\item{x}{(\code{tbl_svysummary})\cr
Object of class \code{'tbl_svysummary'} typically created with \code{gtsummary::tbl_svysummary()}.}

\item{header}{(\code{string})\cr
String specifying updated header.
Review \code{gtsummary::modify_header()} for details on use.}
}
\value{
a 'gtsummary' table
}
\description{
Add ESS Header to gtsummary tbl
}
\examples{
\dontshow{if (rlang::is_installed(c("survey", "gtsummary", "cards", "cardx", "dplyr", "tidyr"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
svy <- survey::svydesign(~1, data = gtsummary::trial, weights = ~1)

gtsummary::tbl_svysummary(svy, include = age) |>
  add_ess_header()

gtsummary::tbl_svysummary(svy, by = trt, include = age) |>
  add_ess_header(header = "**{level}**  \nN = {n} / {N}")
\dontshow{\}) # examplesIf}
}
