% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ess_header.R
\name{add_ess_header}
\alias{add_ess_header}
\title{Add ESS Table Header}
\usage{
add_ess_header(x, header = "**{level}**  \\nESS = {round(n, 1)}")
}
\arguments{
\item{x}{(\code{tbl_svysummary})\cr
Object of class \code{'tbl_svysummary'} typically created with \code{gtsummary::tbl_svysummary()}.}

\item{header}{(\code{string})\cr
String specifying updated header.
Review \code{gtsummary::modify_header()} for details on use.}
}
\value{
a 'gtsummary' table
}
\description{
This function replaces the counts in the default header of
\code{gtsummary::tbl_svysummary()} tables to counts representing the
Effective Sample Size (ESS). See \code{\link[=ess]{ess()}} for details.
}
\examples{
\dontshow{if (rlang::is_installed(c("survey", "gtsummary", "cards", "cardx", "dplyr"))) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
svy <- survey::svydesign(~1, data = nhefs_weights, weights = ~ w_ate)

gtsummary::tbl_svysummary(svy, include = c(age, sex, smokeyrs)) |>
  add_ess_header()

gtsummary::tbl_svysummary(svy, by = qsmk, include = c(age, sex, smokeyrs)) |>
  add_ess_header(header = "**{level}**  \nN = {round(n, 1)} / {round(N, 1)}")
\dontshow{\}) # examplesIf}
}
