% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-methods.R
\name{plot-halfmoon}
\alias{plot-halfmoon}
\alias{plot.halfmoon_balance}
\alias{plot.halfmoon_ess}
\alias{plot.halfmoon_calibration}
\alias{plot.halfmoon_roc}
\alias{plot.halfmoon_auc}
\alias{plot.halfmoon_qq}
\title{Plot Methods for halfmoon Objects}
\usage{
\method{plot}{halfmoon_balance}(x, ...)

\method{plot}{halfmoon_ess}(x, ...)

\method{plot}{halfmoon_calibration}(x, ...)

\method{plot}{halfmoon_roc}(x, ...)

\method{plot}{halfmoon_auc}(x, ...)

\method{plot}{halfmoon_qq}(x, ...)
}
\arguments{
\item{x}{A halfmoon data object with appropriate class}

\item{...}{Additional arguments passed to autoplot()}
}
\value{
Invisibly returns the ggplot2 object after printing
}
\description{
These methods provide standard plot generation for halfmoon data objects.
They create the plot using autoplot() and then print it.
}
