% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyselect-reexports.R, R/tidysmd-rexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{peek_vars}
\alias{select_helpers}
\alias{contains}
\alias{ends_with}
\alias{everything}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{starts_with}
\alias{last_col}
\alias{geom_love}
\alias{tidysmd}
\alias{love_plot}
\alias{tidy_smd}
\alias{bind_matches}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{tidyselect}{\code{\link[tidyselect:starts_with]{contains}}, \code{\link[tidyselect:starts_with]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect:everything]{last_col}}, \code{\link[tidyselect:starts_with]{matches}}, \code{\link[tidyselect:starts_with]{num_range}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{peek_vars}}, \code{\link[tidyselect]{starts_with}}}

  \item{tidysmd}{\code{\link[tidysmd]{bind_matches}}, \code{\link[tidysmd]{geom_love}}, \code{\link[tidysmd:geom_love]{love_plot}}, \code{\link[tidysmd]{tidy_smd}}}
}}

