% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_roc.R
\docType{data}
\name{stat_roc}
\alias{stat_roc}
\alias{StatRoc}
\title{ROC Curve Stat}
\usage{
stat_roc(
  mapping = NULL,
  data = NULL,
  geom = "path",
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  .focal_level = NULL,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{aes()}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot.}

\item{data}{The data to be displayed in this layer. If \code{NULL}, the default,
the data is inherited from the plot data as specified in the call to \code{ggplot()}.}

\item{geom}{Geometric object to use. Default is "path".}

\item{position}{Position adjustment.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with a warning.
If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{Logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather than
combining with them.}

\item{.focal_level}{The level of the outcome variable to consider as the
treatment/event. If \code{NULL} (default), uses the last level for factors or
the maximum value for numeric variables.}

\item{...}{Other arguments passed on to layer().}
}
\value{
A ggplot2 layer.
}
\description{
Statistical transformation for ROC curves.
}
\keyword{datasets}
