% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\docType{data}
\name{stat_roc}
\alias{stat_roc}
\alias{StatRoc}
\title{ROC Curve Stat}
\usage{
stat_roc(
  mapping = NULL,
  data = NULL,
  geom = "path",
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings.}

\item{data}{Data frame.}

\item{geom}{Geometric object to use. Default is "path".}

\item{position}{Position adjustment.}

\item{na.rm}{Remove missing values? Default TRUE.}

\item{show.legend}{Show legend? Default NA.}

\item{inherit.aes}{Inherit aesthetics? Default TRUE.}

\item{...}{Additional arguments.}
}
\value{
A ggplot2 layer.
}
\description{
Statistical transformation for ROC curves.
}
\keyword{datasets}
